+++ pkg/vm/lib/transformations/type_flow/unboxing_info.dart
@@ -176,6 +176,14 @@ class UnboxingInfoManager {
       final inferredType = typeFlowAnalysis.getFieldValue(member).value;
       if (member.hasSetter) {
         _applyToArg(member, unboxingInfo, 0, inferredType);
+        // Arguments of implicit setters for covariant fields
+        // cannot be unboxed based on the field type as setter
+        // performs a type check before value is assigned to the field.
+        if (member.isCovariantByDeclaration) {
+          unboxingInfo.argsInfo.length = 0;
+        } else {
+          _applyToArg(member, unboxingInfo, 0, inferredType);
+        }
       }
       _applyToReturn(member, unboxingInfo, inferredType);
     } else {