+++ sdk/lib/_internal/wasm/lib/js_interop_patch.dart
@@ -545,20 +545,18 @@ extension JSAnyOperatorExtension on JSAny? {
       '(o, a) => o || a', this.toExternRef, any.toExternRef)) as JSAny?;
 
   @patch
-  bool get not => JSBoolean._(
-          JSValue(js_helper.JS<WasmExternRef?>('(o) => !o', this.toExternRef)))
-      .toDart;
+  JSBoolean get not => JSBoolean._(
+      JSValue(js_helper.JS<WasmExternRef?>('(o) => !o', this.toExternRef)));
 
   @patch
-  bool get isTruthy => JSBoolean._(
-          JSValue(js_helper.JS<WasmExternRef?>('(o) => !!o', this.toExternRef)))
-      .toDart;
+  JSBoolean get isTruthy => JSBoolean._(
+      JSValue(js_helper.JS<WasmExternRef?>('(o) => !!o', this.toExternRef)));
 }
 
 @patch
-JSPromise<JSObject> importModule(String moduleName) =>
+JSPromise<JSObject> importModule(JSAny moduleName) =>
     JSPromise<JSObject>._(JSValue(js_helper.JS<WasmExternRef?>(
-        '(m) => import(m)', moduleName.toJS.toExternRef)));
+        '(m) => import(m)', moduleName.toExternRef)));
 
 @JS('Array')
 @staticInterop