+++ sdk/lib/js_interop/js_interop.dart
@@ -1143,12 +1143,10 @@ extension JSAnyOperatorExtension on JSAny? {
   external JSAny? or(JSAny? any);
 
   /// The result of <code>!`this`</code> in JavaScript.
-  // TODO(srujzs): Change this to JSBoolean to be consistent.
-  external bool get not;
+  external JSBoolean get not;
 
   /// The result of <code>!!`this`</code> in JavaScript.
-  // TODO(srujzs): Change this to JSBoolean to be consistent.
-  external bool get isTruthy;
+  external JSBoolean get isTruthy;
 }
 
 /// The global scope that is used to find user-declared interop members.
@@ -1195,4 +1193,4 @@ external JSObject createJSInteropWrapper<T extends Object>(T dartObject);
 ///
 /// Returns a [JSPromise] that resolves to a [JSObject] that's the module
 /// namespace object.
-external JSPromise<JSObject> importModule(String moduleName);
+external JSPromise<JSObject> importModule(JSAny moduleName);