+++ CHANGELOG.md
@@ -51,6 +51,16 @@
 
 [#53785]: https://github.com/dart-lang/sdk/issues/53785
 
+#### `dart:js_interop`
+
+- **Breaking Change** [#55508][]: `importModule` now accepts a `JSAny` instead
+  of a `String` to support other JS values as well, like `TrustedScriptURL`s.
+- **Breaking Change** [#55267][]: `isTruthy` and `not` now return `JSBoolean`
+  instead of `bool` to be consistent with the other operators.
+
+[#55508]: https://github.com/dart-lang/sdk/issues/55508
+[#55267]: https://github.com/dart-lang/sdk/issues/55267
+
 ### Tools
 
 #### Linter